pyproject.toml
[tool.poetry]
name = "myrtle"
version = "0.1.0"
description = ""
authors = ["Jonas Stettner <mail@jstet.net>"]
readme = "README.md"
packages = [{include = "myrtle"}]

[tool.poetry.scripts]
start = "myrtle.main:start"

[tool.poetry.dependencies]
python = "^3.11"
dinosay = "^1.2.0"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
modal = "^0.50.2614"
huggingface-hub = "^0.16.4"
rich = "^13.4.2"
requests = "^2.31.0"
pandas = "^2.0.3"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.21.1"
jupyter = "^1.0.0"
contractions = "^0.1.73"
autocorrect = "^2.6.1"
datasets = "^2.13.1"
transformers = "^4.30.2"
torch = "2.0.0"
accelerate = "^0.20.3"
python-dotenv = "^1.0.0"
langdetect = "^1.0.9"
language-tool-python = "^2.7.1"
alt-profanity-check = "^1.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.277"
pre-commit = "^3.3.3"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
importlib-resources = "^6.0.0"
faker = "^19.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py311']
preview = true
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    "tests/data",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


# Assume Python 3.10.
target-version = "py311"